simplify postgres slt, create individual converter for polygon, create helper file for converter functions
