+++ java/connector-node/risingwave-connector-service/src/main/resources/postgres.properties
@@ -1,7 +1,8 @@
 # configs for postgres conneoctor
 connector.class=io.debezium.connector.postgresql.PostgresConnector
-converters=textrepr,vec2str
-textrepr.type=com.risingwave.connector.cdc.debezium.converters.PgTextRepresentationConverter
+converters=st2str,polygon2str,vec2str
+st2str.type=com.risingwave.connector.cdc.debezium.converters.PgCompositeToStringConverter
+polygon2str.type=com.risingwave.connector.cdc.debezium.converters.PgPolygonToStringConverter
 vec2str.type=com.risingwave.connector.cdc.debezium.converters.PgVectorToStringConverter
 # default snapshot mode to initial
 snapshot.mode=${debezium.snapshot.mode:-initial}
