+++ java/connector-node/risingwave-source-cdc/src/main/java/com/risingwave/connector/cdc/debezium/converters/PgVectorToStringConverter.java
@@ -18,8 +18,6 @@ package com.risingwave.connector.cdc.debezium.converters;
 
 import io.debezium.spi.converter.CustomConverter;
 import io.debezium.spi.converter.RelationalColumn;
-import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
 import java.util.Properties;
 import org.apache.kafka.connect.data.SchemaBuilder;
 
@@ -40,16 +38,6 @@ import org.apache.kafka.connect.data.SchemaBuilder;
  */
 public class PgVectorToStringConverter implements CustomConverter<SchemaBuilder, RelationalColumn> {
 
-    // Must match `DEBEZIUM_UNAVAILABLE_VALUE` on the Rust side
-    // (src/common/src/types/mod.rs). When a vector column is TOAST'd and the
-    // UPDATE does not touch it, Debezium hands us a bare `java.lang.Object`
-    // singleton instead of the usual placeholder string, because the default
-    // schema for vector is ARRAY (a String placeholder would fail schema
-    // validation). We translate that sentinel back to the canonical string so
-    // the downstream RisingWave parser + materialize executor handle it
-    // through the same path as varchar/jsonb/bytea TOAST columns.
-    private static final String UNAVAILABLE_VALUE_PLACEHOLDER = "__debezium_unavailable_value";
-
     @Override
     public void configure(Properties props) {}
 
@@ -58,31 +46,10 @@ public class PgVectorToStringConverter implements CustomConverter<SchemaBuilder,
             RelationalColumn column, ConverterRegistration<SchemaBuilder> registration) {
         // pgvector ships the `vector` type as an extension; the OID is assigned dynamically at
         // CREATE EXTENSION time, so we match on the type name instead.
-        if ("vector".equalsIgnoreCase(column.typeName())) {
-            registration.register(SchemaBuilder.string().optional(), this::convertVector);
-        }
-    }
 
-    private String convertVector(Object value) {
-        if (value == null) {
-            return null;
+        if ("vector".equalsIgnoreCase(column.typeName())) {
+            registration.register(
+                    SchemaBuilder.string().optional(), PgTextConversionUtils::convertScalar);
         }
-        // Debezium's unchanged-TOAST sentinel for non-STRING schema columns is a
-        // bare `java.lang.Object` singleton. Normal vector values arrive as a
-        // concrete subclass (String / byte[] / PGobject / ...), so a value whose
-        // runtime class is exactly Object can only be the sentinel.
-        if (value.getClass() == Object.class) {
-            return UNAVAILABLE_VALUE_PLACEHOLDER;
-        }
-        if (value instanceof byte[] bytes) {
-            return new String(bytes, StandardCharsets.UTF_8);
-        }
-        if (value instanceof ByteBuffer buffer) {
-            var readOnly = buffer.asReadOnlyBuffer();
-            var bytes = new byte[readOnly.remaining()];
-            readOnly.get(bytes);
-            return new String(bytes, StandardCharsets.UTF_8);
-        }
-        return value.toString();
     }
 }
